requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
certifi==2024.2.2
werkzeug==3.0.1
email-validator==2.1.0.post1
PyJWT==2.8.0
python-multipart==0.0.6
openai==1.3.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
pytest==6.2.5
pytest-cov==2.12.1
redis==5.0.1 